Add TESTING.md and deployment evidence
This commit adds a new TESTING.md file with associated file structure in
the project.

The first entry to the TESTING.md file is evidence of successful
deployment to Heroku, where it prints out some test data from the
connected google sheets document to the console on the deployed Heroku
site.