Add basic menu functionality for trading simulator data
This commit adds a basic initial menu to allow for the user to perform
the simulation of the trading app generating the end of day trading
file.

It currently gets saved into a folder as part of the project, but will
be transferred to an output folder as part of the googledrive project.