Add basic menu logic and data transfer functionality
This commit adds some additional basic menu functionality to test the
application menu flow.

It also adds the functionality to transfer the data from the trading app
output file onto the FX Net database trades worksheet, ready for later
analysis.

A helper function to roll the system date was created in order to ensure
that a simulated trade file is never created for the same trade date.