Refactor menus using InquirerPy and edit trade_file_generation function
This commit refactors one of the menus using InquirerPy as this library
has good error control for inputs, but using choices instead of typing.

The trade_file_generation function has also been edited to output a file
name along with the data. This is so it can be passed into the
create_output_file function when it saves it to google drive. Removing
the need to calculate the file name.