Refactor of helper functions for managing googledrive files
This commit has a refactor of the helper function that shows a list of files
in google drive. It now adds the functionality to filter out the files you want to see.
This was added so that when generating all the test data from the Trading App, it can be
easily deleted by calling the delete_file function.