Add functions for checking available files for loading
This commit adds 3 functions that assist with checking if files are eligible for loading into
FX Net Application.

One checks what files with prefix "trade_data" is in the shared location.
One checks what files have been added to the FILES_LOADED table in FX Net.
The third compares these and returns a list of eligble for loading files.