Add functionality to select file to load into FX Net
This commit adds the functionality for the user to select a file they want to load
into FX Net. Previously, the file_id was passed directly into FX_Net after generation,
but as these are two separate applications it makes sense to separate functionality.
It now shows the user which files are available to load, by comparing what has already
been loaded vs what it available to load.