Add feature to show console netting summary by value date
This commit adds the functionality for the user to select the value date in which they want to
produce the netting report for.
It checks the TRADES table and creates a list of unique value dates for the user to select.
Previously it was running for the most recent date, this change gives more flexibility and
control to the user.