Add prototype of new menu function call to test.py
This commit doesn't change production code, but adds a prototype of the menu
system in test.py.

Instead of reading in a question string and a choices list, it will now read in a
question string and a choices dictionary. The dictionary will contain the choice
answer as before but also be directly linked to the relevant action/function
based on that choice.

The function returns the executed function.