Add functionality to menu system in the test.py prototype
This commit enhances the previous menu prototype. It previously was enhanced to
take in a dictionary of choices and functions, and had a separate question string.
This commit changes so the entire "question, answer and function" are all part of
one dictionary. Reducing the amount of parameters to load into menu function.