Set up folder/file restructuring
This commit adds a new file structure to organise the code in a cleaner way.

A lib folder to contain all project code
A folder for app_selector, trading_simulator and fx_net.
An init file to ensure correct linkage of the files.
A utils.py file to house shared resources.
And a test in a main.py file to ensure all the files are linked properly. (will be removed)