Refactor of fuzzy_select_menu and formated some print statements
This commit refactoered the fuzzy_select_menu into the utils module.
I has the ability to validate the choice against the intial list
of choices and shows an error to the user on the screen.
Functionality part of the InquirerPy library.

Specific peice of documentation implements a more complex version
here, https://inquirerpy.readthedocs.io/en/latest/pages/prompts/fuzzy.html#codecell1
and my code is a trimmed down version of this only using the validate and
invalid_message parts.