Testing feature to add filter to trade count by client and getting dates list from trade table
This commit adds two features. One to the trade_count_by_client function. It now allows for it to be passed in the True bool if the function is to be used to run on one specific trade date.
It utilises code structure defined in the netting_summary_by_value_date where the get_available_report_value_dates function length is checked and returns if there is no data available.
It then checks if the bool flag is set to true and if so will initialise the fuzzy_select_menu from utils and setting the date variable, using this to filter the pandas data frame.

The helper function of get_available_report_value_dates has been also refactored to allow for selection of trade of value, so it can serve multiple uses.

**Potential further refactor is to define the table names into a constant variable file so that these are not hard coded into the functions themselves. Allowing for better re-usability.