Add to feature create_netting_report_by_value_date
This commit starts the feature of creating the netting report by value date within the function create_netting_report_by_value_date.

It utilises the helper functions get_available_report_dates_by_type and fuzzy_select_menu, to get available report dates that can be passed into the fuzzy selection menu type.
Checking if the list is blank first and running if not.
It creates a new file using the google drive api and shares with an email so its visible.

This specific commit initially adds the supporting trade data, and will also contain the netting breakdown with the next feature update.