Add the base funtionality to create_payment_files
This commit adds the beginning of the functionality to create_payment_files. It has the ability to generate an output file and interogates the fx_net_db database (workbook) and returns the data.

A need addition to the database.py file connects and creates a new variable for the table for payment instructions.