Add remaining logic to create_payment_files in test.py
This commit adds the remaining logic to create_payment_files in the test.py file before moving to production.
It utilises pandas again to create dataframes and merge them based on matching client and ccy fields. Creating a new datafrme containing the client, ccy amount and payment details.
It then filters only for the negative numbers which are the outgoing payments that need to be sent to the downstream payment system.