Add testing for empty database scenarios. Add helper function to tested functions
This commit tests the empty database scenarios withing the reporting menu context. Checking to see if the relevant user message appears if there is no data in the TRADES table.

I identified an opportunity for refactoring the display message into its own function. Which I added into the logic of the respective functions and moved no_data_message into the utils.py module.